requirements.txt
# Core - Agent
anthropic>=0.42.0
openai>=1.0
pyyaml>=6.0

# Token counting
tiktoken>=0.8.0

# CLI interface
rich>=13.0
prompt-toolkit>=3.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24.0

# Utilities
python-dotenv>=1.0

# Ingestion
pypdf>=4.0
# openai-whisper>=20231117  # optional, uncomment for audio transcription
